### 2017-02-19 #183 @arfon
:Wave @noamross - this package looks like a potential good candidate for rOpenSci because of the focus on data visualization.
