### 2017-02-20 06b3c09
update readme for increased clarity
