### 2017-02-20 db7f3b3
extended readme with additional intuition, comments on tests, and directions on contributions, problems, questions, and requests.
