+++ README.md
@@ -8,7 +8,9 @@ plotrr helps address this issue by providing several functions that make visual 
 
 As demonstrated in Crabtree and Nelson (2017), creating and interpreting plots like this this can help scholars find initial support for their theoretical expectations prior to conducting analysis with pooled data. The intuition here is that researchers can check their initial priors about relationships within cases. When the data support those priors, scholars have some additional evidence that the processes they theorize actually occur in the real world.
 
-Another set of functions helps the researcher evaluate correlations between measures within groups/units. `groupcor` returns a tibble data frame with group/unit identifiers and the correlation coefficient between two measures for each group/unit in a data frame. `groupcorplot` performs a similar function and returns a figure that plots the correlation coefficient between measures within groups/units. These functions can help researchers identify important patterns across groups/units.
+Another set of functions helps the researcher evaluate correlations between measures within groups/units. `groupcor` returns a tibble data frame with group/unit identifiers and the correlation coefficient between two measures for each group/unit in a data frame. `groupcorplot` performs a similar function and returns a figure that plots the correlation coefficient between measures within groups/units. These functions can help researchers identify important patterns across groups/units. 
+
+They can also help scholars check the correlation between competing measures - such as different measures of state human rights practices - within units. This is one way to check the construct validity of the measures (Trochim & Donnelly, 2008). This also might be useful to generate new empirical puzzles; if measures disagree in important cases, why is this the case? 
 
 In addition to these functions, the package also includes `histplots`, which creates histograms of a measure for each group/unit, and `bivarrugplot`, which returns a plot of the bivariate relationship between two measures alongside a rugplot of each measure.
 
@@ -31,8 +33,14 @@ install.packages("plotrr")
 ```
 
 ## Support or Contact
-Having trouble with plotrr? Please email [Charles Crabtree](mailto:ccrabtr@umich.edu) and he'll help you sort it out.
+Please use the issue tracker for problems, questions, or feature requests. If you would rather email, you can contact [Charles Crabtree](mailto:ccrabtr@umich.edu) and he will address the issue.
+
+If you would like to contribute to the package, that is great! We welcome pull requests and new developers.
+
+## Tests
+To test the software, users and potential contributors can use the example code provided in the documentation for each function.
 
 ### References
 - Crabtree, Charles, and Michael J. Nelson. 2017. "New Evidence for a Positive Relationship Between De Facto Judicial Independence and State Respect for Empowerment Rights." _International Studies Quarterly_.
+- Trochim, William M. K., and James P. Donnelly. 2008. _Research Methods Knowledge Base_. New York, NY: Cengage Learning.
 - Tukey, John W. 1977. _Exploratory Data Analysis_. New York, NY: Pearson.
