### 2017-02-27 #183 @noamross
@arfon Thanks. This doesn't meet the threshold for us as an extension beyond base or ggplot graphics.
