### 2017-02-28 #190 @whedon
[REVIEW]: plotrr: Functions for making visual exploratory data analysis with nested data easier. Edit
**Submitting author:** @cdcrabtree (<a href="http://orcid.org/0000-0001-5144-8671">Charles Crabtree</a>)
**Repository:** <a href="https://github.com/cdcrabtree/plotrr" target ="_blank">https://github.com/cdcrabtree/plotrr</a>
**Version:** v0.2.0
