### 2017-03-01 ec3d0d2
Calls to aes() and aes_string() need ggplot2::
