### 2017-03-01 c5c8c27
length.unique: export as length.unique, not S3 method
