+++ NAMESPACE
@@ -1,6 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(length,unique)
 export(bivarplots)
 export(bivarrugplot)
 export(clear)
@@ -8,6 +7,7 @@ export(dotplots)
 export(groupcor)
 export(groupcorplot)
 export(histplots)
+export(length.unique)
 export(makefacnum)
 export(violinplots)
 importFrom(dplyr,"%>%")
