+++ NAMESPACE
@@ -11,4 +11,5 @@ export(length.unique)
 export(makefacnum)
 export(violinplots)
 importFrom(dplyr,"%>%")
+importFrom(dplyr,"group_by")
 importFrom(stats,"cor")
