### 2017-03-01 5ca5018
examples for groupcor and groupcorplot: remove print(dat)
