+++ R/groupcorplot.r
@@ -15,7 +15,6 @@
 #'c <- rep(c(1:10), times = 100)
 #'data <- data.frame(a, b, c)
 #'groupcorplot("a", "b", "c", data, "pearson")
-#'print(dat)
 #'}
 #'@importFrom dplyr "%>%" "group_by"
 #'@importFrom stats "cor"
