### 2017-03-01 #190 @kbroman
I sent @cdcrabtree a pull request with several small changes, plus added a couple of issues with a few other minor suggestions.

Removing `\dontrun{}` from the examples would make it so `R CMD check` would give at least the simplest tests, of whether the functions can run without giving an error.

It would be great if the authors could add a formal [vignette](http://kbroman.org/pkg_primer/pages/vignettes.html) to give a bit more detail on the use of the functions, and maybe some discussion of interpretation. But the ReadMe file and the examples in the function documentation are probably sufficient.

The [CRAN](https://cran.r-project.org) requirement on providing license information within an R package is a bit odd, in that the `LICENSE` file just says year and copyright holder and doesn't mention the license. It might be good to add another `License.md` file that is more explicit about the MIT license, and then add that to `.Rbuildignore` so that it doesn't get included in the package file sent to CRAN.
