### 2017-03-01 e963c63
changes recommended by kbroman
