+++ R/bivarplots.r
@@ -8,13 +8,11 @@
 #'@param data A data frame.
 #'@return A series of figures that plot the bivariate relationship between two measures for each group/unit.
 #'@examples
-#'\dontrun{
 #'a <- runif(1000, min = 0, max = 1)
 #'b <- a + rnorm(1000, mean = 0, sd =1)
 #'c <- rep(c(1:10), times = 100)
 #'data <- data.frame(a, b, c)
 #'bivarplots("a", "b", "c", data)
-#'}
 #'@export
 
 bivarplots <- function(x, y, group, data) {
