+++ man/lengthunique.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/lengthunique.r
-\name{length.unique}
-\alias{length.unique}
+\name{lengthunique}
+\alias{lengthunique}
 \title{Calculates the number of unique values in a vector}
 \usage{
-\method{length}{unique}(x)
+lengthunique(x)
 }
 \arguments{
 \item{x}{A vector.}
@@ -24,4 +24,3 @@ length.unique(x)
 \author{
 Charles Crabtree \email{ccrabtr@umich.edu}
 }
-
