+++ man/dotplots.Rd
@@ -35,4 +35,3 @@ dotplots("a", "b", "c", data, 20)
 \author{
 Charles Crabtree \email{ccrabtr@umich.edu}
 }
-
