+++ man/groupcor.Rd
@@ -36,4 +36,3 @@ print(dat)
 \author{
 Charles Crabtree \email{ccrabtr@umich.edu}
 }
-
