### 2017-03-01 cc535b5
updates to address issues #2, #3, and #4 plus other reviewer requests
