+++ R/bivarrugplot.r
@@ -7,12 +7,10 @@
 #'@param data A data frame.
 #'@return A plot of the bivariate relationship between two measures with a rugplot for each measure.
 #'@examples
-#'\dontrun{
 #'a <- runif(1000, min = 0, max = 1)
 #'b <- a + rnorm(1000, mean = 0, sd =1)
 #'data <- data.frame(a, b)
 #'bivarrugplot("a", "b", data)
-#'}
 #'@export
 
 bivarrugplot <- function(x, y, data) {
