+++ R/clear.r
@@ -4,9 +4,7 @@
 #'@author Charles Crabtree \email{ccrabtr@umich.edu}
 #'@param ... An unused argument.
 #'@examples
-#'\dontrun{
 #'clear()
-#'}
 #'@export
 
 clear <- function(...) {
