+++ R/groupcorplot.r
@@ -9,13 +9,11 @@
 #'@param stat The type of correlation coefficient to be calculated. One of "pearson", "kendall", or "spearman".
 #'@return A figure that plots the correlation coefficient between measures across groups/units.
 #'@examples
-#'\dontrun{
 #'a <- runif(1000, min = 0, max = 1)
 #'b <- a + rnorm(1000, mean = 0, sd =1)
 #'c <- rep(c(1:10), times = 100)
 #'data <- data.frame(a, b, c)
 #'groupcorplot("a", "b", "c", data, "pearson")
-#'}
 #'@importFrom dplyr "%>%" "group_by"
 #'@importFrom stats "cor"
 #'@export
