+++ R/lengthunique.r
@@ -5,11 +5,9 @@
 #'@param x A vector.
 #'@return The number of unique values in a vector.
 #'@examples
-#'\dontrun{
 #'x <- rep(c(1:10), 10)
-#'length.unique(x)
-#'}
-#'@export length.unique
+#'lengthunique(x)
+#'@export lengthunique
 
 lengthunique <- function(x) {
   length(unique(x[!is.na(x)]))
