+++ R/histplots.r
@@ -9,13 +9,11 @@
 #'@param n The number of bins.
 #'@return Historgrams for a measure for each group/unit.
 #'@examples
-#'\dontrun{
 #'a <- runif(1000, min = 0, max = 1)
 #'b <- a + rnorm(1000, mean = 0, sd =1)
 #'c <- rep(c(1:10), times = 100)
 #'data <- data.frame(a, b, c)
 #'histplots("a", "b", "c", data, 5)
-#'}
 #'@export
 
 histplots <- function(x, y, group, data, n) {
