+++ R/dotplots.r
@@ -9,13 +9,11 @@
 #'@param n The number of bins. Some experimentation with this number might be necessary.
 #'@return Historgrams for a measure for each group/unit.
 #'@examples
-#'\dontrun{
 #'a <- runif(1000, min = 0, max = 1)
 #'b <- a + rnorm(1000, mean = 0, sd =1)
 #'c <- rep(c(1:10), times = 100)
 #'data <- data.frame(a, b, c)
 #'dotplots("a", "b", "c", data, 20)
-#'}
 #'@export
 
 dotplots <- function(x, y, group, data, n) {
