+++ R/makefacnum.r
@@ -5,14 +5,12 @@
 #'@param x A vector.
 #'@return A numeric vector.
 #'@examples
-#'\dontrun{
 #'x <- c("x", "y", "z")
 #'x <- as.factor(x)
 #'x
 #'x <- makefacnum(x)
 #'x
 #'is.numeric(x)
-#'}
 #'@export
 
 makefacnum <- function(x) {
