### 2017-03-01 f361c47
add vignette and make some minor fixes
