+++ R/bivarrugplot.r
@@ -8,7 +8,7 @@
 #'@return A plot of the bivariate relationship between two measures with a rugplot for each measure.
 #'@examples
 #'a <- runif(1000, min = 0, max = 1)
-#'b <- a + rnorm(1000, mean = 0, sd =1)
+#'b <- a + rnorm(1000, mean = 0, sd = 1)
 #'data <- data.frame(a, b)
 #'bivarrugplot("a", "b", data)
 #'@export
