+++ vignettes/plotrr-vignette.Rmd
@@ -1,0 +1,97 @@
+---
+title: "Vignette for plotrr"
+author: "Charles Crabtree and Michael J. Nelson"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Vignette for plotrr}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+## plotrr: Functions for making visual exploratory data analysis with nested data easier.
+
+Social scientists can improve their research by conducting exploratory data analysis (EDA) (Tukey 1977). The benefits of EDA include: ''maximiz[ing] insight into a data set; uncover[ing] underlying structure; extract[ing] important variables; detect[ing] outliers and anomalies; test[ing] underlying assumptions; develop[ing] parsimonious models; and determin[ing] optimal factor settings'' (NIST/SEMATECH 2012). Despite these benefits, scholars infrequently conduct EDA. One possible explanation for this is because it takes additional time to do so; it is often easier to move straight to confirmatory analysis.  
+
+The time concern is particularly an issue for researchers who use nested data. The issue here is that most existing EDA software routines visualize relationships based on the pooled data. Few existing functions help scholars _easily_ visualize relationships within groups/units. 
+
+plotrr helps address this gap in existing software by providing several functions that make visual EDA easier to conduct.  The focus of many of the package's functions is to create plots that can help researchers explore relationships within nested data. Among other things, these functions can help scholars assess the extent to which expected relationships between variables occur in specific cases. 
+
+In the following sections, we provide brief examples of the core functions in the package. Since we use simulated data in the examples, it is difficult to interpert the visual results in a meaningful way. In evaluating the outputs of these functions, researchers will need to rely on their understanding of the data and existing theoretical assumptions.
+
+### Functions that illustrate the relationship between variables within groups/units
+
+One set of functions creates plots that illustrate the relationship between variables within groups/units. As demonstrated in Crabtree and Nelson (2017), creating and interpreting plots like this can help scholars find initial support for their theoretical expectations prior to conducting analysis with pooled data. The intuition here is that researchers can check their initial priors about relationships within cases. When the data support those priors, scholars have some additional evidence that the processes they theorize actually occur in the real world.
+
+#### `bivarplots` 
+This function creates a bivariate plot for every group/unit in the data. Please keep in mind the number of groups/units in your data before running this and other visualization functions in this package. One way to check the number of units is with the included `lengthunique` function.
+
+```{r, fig.show='hold'}
+library(plotrr)
+a <- runif(400, min = 0, max = 1)
+b <- a + rnorm(400, mean = 0, sd = 1)
+c <- rep(c(1:4), times = 100)
+data <- data.frame(a, b, c)
+bivarplots("a", "b", "c", data)
+```
+
+#### `dotplots` 
+This function creates a dot plot for every group/unit in the data.
+```{r, fig.show='hold'}
+library(plotrr)
+a <- runif(400, min = 0, max = 1)
+b <- a + rnorm(400, mean = 0, sd = 1)
+c <- rep(c(1:4), times = 100)
+data <- data.frame(a, b, c)
+dotplots("a", "b", "c", data, 20)
+```
+
+#### `violinplots` 
+This function creates a violin plot for every group/unit in the data.
+```{r, fig.show='hold'}
+library(plotrr)
+a <- runif(400, min = 0, max = 1)
+b <- a + rnorm(400, mean = 0, sd = 1)
+c <- rep(c(1:4), times = 100)
+data <- data.frame(a, b, c)
+violinplots("a", "b", "c", data)
+```
+
[…]
