### 2017-03-01 #190 @cdcrabtree
@kbroman Thanks for the helpful suggestions and changes. We've just finished going through and incorporating your pull request, addressing the issues, and addressing your comments here. We've updated the version number of the software to reflect these changes and creates a NEWS file that enumerates them. Specific to your comments here, we've removed all instances of `dontrun{}`, added a vignette that focuses on the core functions and explains a bit more the intuition behind why researchers might want to use them, and created a `License.md` file. Finally, we've added a note to the README that acknowledges your help with this package.

One thing that I wanted to add here is that our references do not have DOIs, unfortunately. Several are books, which were not assigned a DOI. The article should have one, but the publisher hasn't assigned one yet.

Please let us know if there are any other issues that we can address.

Thanks again for the very helpful comments.
