### 2017-03-01 e5580f8
minor readme update
