### 2017-03-01 34cb816
Small change in ReadMe (link to @kbroman in "Thanks")
