### 2017-03-01 e1c1a8f
updated intro paragraphs with language from vignette and included a reference
