### 2017-03-05 #190 @cdcrabtree
@arfon Here is the Zenodo DOI: 10.5281/zenodo.344900. We plan on making future updates to this package. Is there a way of updating the DOI as we move forward? Or is it always pegged to the reviewed software release? Thanks!
