### 2017-03-05 #190 @pjotrp
It is pegged. One advantage is that you can submit a new paper on a future release.
