### 2017-03-05 #190 @sagetrey
Unsubscribe

Sent from my iPhone

> On Mar 5, 2017, at 4:49 AM, Arfon Smith <notifications@github.com> wrote:
>
> @cdcrabtree - could you move the references you currently have in the paper.md file into a paper.bib file and cite them directly please? (You can read how to do that here)
>
