+++ paper/paper.bib
@@ -1,0 +1,33 @@
+%% This BibTeX bibliography file was created using BibDesk.
+%% http://bibdesk.sourceforge.net/
+
+%% Created for Charles Crabtree at 2017-03-05 08:27:47 -0700 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
+@article{CrabtreeNelson2017,
+	Author = {Charles Crabtree, and Michael J. Nelson},
+	Date-Added = {2017-03-05 15:25:17 +0000},
+	Date-Modified = {2017-03-05 15:25:56 +0000},
+	Journal = {International Studies Quarterly},
+	Title = {New Evidence for a Positive Relationship Between De Facto Judicial Independence and State Respect for Empowerment Rights},
+	Year = {2017}}
+
+@book{Tukey1977,
+	Author = {Tukey, John W.},
+	Date-Added = {2017-03-05 15:23:42 +0000},
+	Date-Modified = {2017-03-05 15:24:07 +0000},
+	Publisher = {Pearson},
+	Title = {Exploratory Data Analysis},
+	Year = {1977}}
+
+@book{TrochimDonnelly2008,
+	Author = {Trochim, William M. K., and James P. Donnelly},
+	Date-Added = {2017-03-05 15:22:58 +0000},
+	Date-Modified = {2017-03-05 15:23:40 +0000},
+	Publisher = {Cengage Learning},
+	Title = {Research Methods Knowledge Base},
+	Year = {2008}}
