### 2017-03-06 #190 @arfon
Thanks for your review @kbroman and editing this submission @pjotrp ✨
@cdcrabtree your paper is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00190 ⚡️ 🚀 💥
