### 2017-03-06 #190 @arfon
> @arfon, if I can help by reviewing something in the future, please let me know. Thanks.

Thanks @cdcrabtree. Please could you add yourself to this file? That way we'll know to call on you sometime in the future: https://github.com/openjournals/joss/blob/master/docs/reviewers.csv
